Git Release with Assets
This PR moves the actual publishing of the GitHub release to after the release assets have been uploaded.

This is important because Github Action triggers a release on publish, so it is possible that release assets might not be visible to the action, as they are yet to be uploaded.

see #https://github.com/release-it/release-it/issues/495