Set `latestTag` and `tagName` on global config in test harness.
As of release-it@13.5.3 `latestTag` and `tagName` are available on the
global configuration object (aka `this.config` within a plugin
instance). These new global config context values make authoring plugins
that care about these values (but don't "control" them) much easier.

Unfortunately, prior to these changes they were not setup and available
from within the test harness. Since these values are always present when
used outside of the test harness having them missing while testing means
that plugins are forced to handle fallback scenarios for when they are
not set (which is extremely unlikely to happen in reality).

This commit sets `latestTag` and `tagName` in the test harness. It
currently uses the values for `latestVersion` and `version` for that
(because the test harness does not actually support providing custom
global configuration to set `git.tagName` template).