Fixes github.proxy not working (#863) (#893)
Based on a recent change in Octokits docs[1], in order for it to use a
configured proxy you must pass another request.agent into the Octokit
constructor. This change adds the `proxy-agent` lib as suggested by
Octokits docs.

Note that with this implmentation, the `http_proxy` or `https_proxy`
environment variables alone will NOT work. Instead one must explicitly
set the value in the `github.proxy` option passed to release-it. This is
to stay inline with current behavior[2].

[1] https://github.com/octokit/octokit.js/commit/2090c85b5addf33b93dac74930312593622c08b8
[2] https://github.com/release-it/release-it/blob/14.14.2/docs/github-releases.md#proxy