Fix `npm.isCollaborator()` on npm v9
With v9.0.0 release [npm][] has changed the CLI commands for package
access. While previously it was `npm access ls-collaborators` and the
output was a JSON object, now it is `npm access list collaborators` and
the output is plain text unless `--json` flag is provided.

This commit adds support for npm v9.0.0 while keeping support for older
npm versions.

[npm]: https://github.com/npm/cli/commit/9c32c6c8d6fc5bdfd6af685731fe26920d7e5446