fix(gitlab): don't set Content-Type header when uploading release assets
As noted in [a comment on #1172][1], switching to the use of the native
fetch API broken uploading release assets to GitLab by hard-coding the
Content-Type header to "text/plain". The [recommended approach][2] is to
avoid setting the Content-Type header at all and let the fetch API
implementation set it so as to ensure that the correct boundary
expression is included.

Add a flag to the `request` function to allow explicitly disabling
setting the Content-Type header. The default behaviour is to set the
header as before but we disable setting the header only in the
`uploadAsset` function.

[1]: https://github.com/release-it/release-it/issues/1172#issuecomment-2468592425
[2]: https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest_API/Using_FormData_Objects#sending_files_using_a_formdata_object

Closes #1172